requirements.txt
google-genai==1.12.1
python-dotenv==1.1.0
requests==2.32.4
bs4==0.0.2
typing_extensions==4.14.0
